PP-13896 - Move custom branding tests into it's own file
- Means we can reuse the tests when using Cypress to check
  the custom branding when ENABLE_REBRAND=true
